requirements.txt
flask
flask_cors
mysql-connector
PyJWT